Use `Test.Utils` for package tests
When we first pulled these packages out, we had to switch from
`Test.Utils.equal` to `Test.Unit.Assert.equal`, because we no longer had
access to the former. With the `test-utils` package now added, we can
switch back to it.

The one exception is our tests in the `date-time` package. This package
only defines helpers for `DateTime`, it doesn't define the `DateTime`
type, so we can't define the `Debug` instance there. That means we have
to define the instance in `cn-debug`. With `cn-debug` relying on
`date-time`, and `test-utils` relying on `cn-debug`, we can't pull
`test-utils` into `date-time` as that would create a cyclical
dependency, so we have to do without the nicer `Test.Utils.equal`
output. This isn't a huge loss, since the advantages of
`Test.Utils.equal` really show with larger or more complex data types,
which we aren't testing here.