Explicitly list all spago dependencies for libs
When we started using the nix-provided version of spago, Dave noticed
that it didn't match the one we get from npm. The nix-provided one is
newer and throws errors if all dependencies aren't explicitly specified.

We fix the issue by adding all of the missing dependencies to the
appropriate `spago.dhall` files.

We can follow up by bumping the npm-provided version of spago in a
future PR.