Change the linting source glob for all packages that have a `src` folder
Following the previous commit, we fix the source glob of the linting
rules for all packages that have their source code under a `src` folder.
This should prevent errors arising from extraneous `*.purs` that should
not be linted.

We also update the `How to: Add a new local package` docs.